backend/requirements.txt
Django>=5.0
channels>=4.0.0
daphne>=4.0.0
djangorestframework>=3.14
django-cors-headers>=4.3
djoser>=2.2
django-allauth==65.13.1
dj-rest-auth==7.0.2
dj-database-url>=2.1
psycopg2-binary>=2.9
Pillow>=10.0
